+++ com.unity.ml-agents/Tests/Runtime/RuntimeExampleTest.cs
@@ -1,0 +1,13 @@
+using NUnit.Framework;
+
+namespace Tests
+{
+    internal class RuntimeExampleTest
+    {
+        [Test]
+        public void RuntimeTestMath()
+        {
+            Assert.AreEqual(2, 1 + 1);
+        }
+    }
+}