+++ com.unity.ml-agents/CHANGELOG.md
@@ -8,11 +8,16 @@ and this project adheres to
 
 ## [Unreleased]
 ### Major Changes
-#### com.unity.ml-agents / com.unity.ml-agents.extensions (C#)
+#### com.unity.ml-agents (C#)
 - Upgraded to Inference Engine 2.2.1 (#6212)
 - The minimum supported Unity version was updated to 6000.0. (#6207)
+- Merged the extension package com.unity.ml-agents.extensions to the main package com.unity.ml-agents. (#6227)
 
 ### Minor Changes
+#### com.unity.ml-agents (C#)
+- Removed broken sample from the package (#6230)
+- Moved to Unity Package documentation as the primary developer documentation. (#6232)
+
 #### ml-agents / ml-agents-envs
 - Bumped grpcio version to >=1.11.0,<=1.53.2 (#6208)